feat: export useAdaptivityWithJSMediaQueries()
