+++ docs/ADAPTIVITY_GUIDE.md
@@ -141,7 +141,7 @@ _Component.module.css_
 создать несколько разметок (для мобильного и десктопа, например) и показывать только одну из них. Для _всплывающих окон_
 такой подход обозначает дублирование атрибута `id`. При этом с точки зрения UX такие окна появляются либо после вызова
 их программно, либо после вызова их пользователем, то есть рендериться на сервере они не будут. Поэтому для _всплывающих
-окон_ мы используем адаптивность через JS. Для этого создан хук [`useAdaptivityWithMediaQueries()`](../src/hooks/useAdaptivityWithMediaQueries.ts).
+окон_ мы используем адаптивность через JS. Для этого создан хук [`useAdaptivityWithJSMediaQueries()`](../src/hooks/useAdaptivityWithJSMediaQueries.ts).
 
 При этом мы помним про показ компонента только после **первого** рендера. Пример:
 
@@ -171,7 +171,7 @@ const App = () => {
 };
 ```
 
-В нашем примере внутри `<ModalRoot />` и `<ModalPage />` используется `useAdaptivityWithMediaQueries()` для
+В нашем примере внутри `<ModalRoot />` и `<ModalPage />` используется `useAdaptivityWithJSMediaQueries()` для
 переключения компонентов на мобильный или десктопный виды.
 
 В библиотеке логику с состоянием `mounted` инкапсулирует в себе компонент [AppRootPortal](../src/components/AppRoot/AppRootPortal.tsx).
